requirements.txt
jenkinsapi
pyaml
